requirements.txt
openai-whisper
sounddevice
pynput
pyperclip
rumps
pyobjc-framework-Cocoa
